feat(S1.5): add MVP inventory foundation and fix agency e2e suite
Lay down seat-assignment concurrency, integer XAF pricing, and v8 schema
before booking APIs so parallel seat holds are safe on PostgreSQL.

+++ src/main/java/cm/jemil/agency/adapter/inbound/rest/AgencyController.java
@@ -90,7 +90,7 @@ public class AgencyController implements AgencyApi {
                 agencyId,
                 addRouteDTO.getOriginCityId(),
                 addRouteDTO.getDestinationCityId(),
-                addRouteDTO.getPrice(),
+                addRouteDTO.getPrice().intValue(),
                 addRouteDTO.getTotalSeats()));
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
